NSIS: Make the licensing page match reality.
The NSIS licensing page is centered around making the user agree to
a EULA. We don't have one of those. Replace the "you must accept" and
"I agree" text with something more neutral.

We could alternatively omit with the licensing page altogether, but
the GPL is an important aspect of the project and deserves some sort
of mention.

Bug: 1115
Change-Id: I5fd5beac2188b61a90c2842de7d7547aa815ff4f
Reviewed-on: https://code.wireshark.org/review/35624
Reviewed-by: Anders Broman <a.broman58@gmail.com>