Netlink: stop hiding packet_netlink_data pointer
The underlying netlink dissector sets up a data structure to help maintain
information about this netlink packet. It gets passed through the familiy
specific netlink dissectors private data structures to reappear when
support funtions of the underlying netlink dissector are needed. In the
mean time a copy of data (the value 'encoding' to be precise) in this
structure is also maintained in these familiy specific netlink dissectors,
adding to the confusion.
This change is to make the underlying netlink dissector data structure
a normal part of the function interfaces, so that it is present without
being dependant on another private data structure.
This change is a first step towards removing the unnessesary copy of the
encoding value.

Change-Id: I69e78a2b15e58e149e82e89c19e519ef041ee6b1
Reviewed-on: https://code.wireshark.org/review/35688
Reviewed-by: Jaap Keuter <jaap.keuter@xs4all.nl>
Petri-Dish: Jaap Keuter <jaap.keuter@xs4all.nl>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>