Netlink: Properly interpret and mask out attribute type
The netlink attribute type is a 16 bit field, of which the two top most
bits are booleans. Interpret them as such.
The remaining 14 bits form the attribute type value. Due to the flexible
way the interpretation is setup, through the use of family specific code,
the header field for the attribute type value has to have a proper mask.
Otherwise the two top bits are taken (incorrectly) as part of the value.
Since this may not be obvious to the netlink family dissector creator
better enforce it by adding the masked value in the underlying netlink
dissector, using whatever header field is given for this.

Change-Id: I791f9b1de01505d4a4793abbcf62e596b864e2f0
Reviewed-on: https://code.wireshark.org/review/35725
Reviewed-by: Jaap Keuter <jaap.keuter@xs4all.nl>
Petri-Dish: Jaap Keuter <jaap.keuter@xs4all.nl>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>