netlink: use value retrieval with proper encoding
This change fixes the problem of value retrieval being done:
1. without regard for the declared encoding,
2. multiple times, for addition to other tree items.

The first issue can result in two different representations of the
same field. The second is an optimization, and a way to make sure
a consistent interpretation of the protocol field is made.

Change-Id: Id521f4deafe381322195b1eb998138efcaa51f30
Reviewed-on: https://code.wireshark.org/review/35745
Reviewed-by: Jaap Keuter <jaap.keuter@xs4all.nl>
Petri-Dish: Jaap Keuter <jaap.keuter@xs4all.nl>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>