netlink: use indicated encoding for attributes
In netlink the general format of attributes includes a network byte order
indication flag in the header. When set this has to be passed to the
attribute value dissector function as the byte ordering fo the attribute
value. Otherwise the heuristically determined capture host byte order has
to be passed to the attribute value dissector function.
With the removal of all copies of 'encoding' this can now be achieved
through setting of the encoding in nl_data.

Change-Id: Iec0c1b2c2958734a469ff6f75db4626846cb30c9
Reviewed-on: https://code.wireshark.org/review/35831
Reviewed-by: Jaap Keuter <jaap.keuter@xs4all.nl>
Petri-Dish: Jaap Keuter <jaap.keuter@xs4all.nl>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>