EVS: Handle A.2.1.3 Special case for 56 bit payload size
The resulting ambiguity between EVS Primary 2.8 kbps and EVS AMR-WB IO
SID frames is resolved through the
most significant bit (MSB) of the first byte of the payload. By
definition, the first data bit d(0) of the EVS Primary 2.8
kbps is always set to 0. Therefore, if the MSB of the first
byte of the payload is set to 0 (see Figure A.2), then the
payload is an EVS Primary 2.8 kbps frame in Compact format. Otherwise it
is an EVS AMR-WB IO SID frame in
Header-Full format with one CMR byte.

Change-Id: I16733698e49ea3651f775b774b59569cfa1c89a1
Reviewed-on: https://code.wireshark.org/review/35976
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>