f5ethtrailer: Whitespace / formatting changes only.
The formatting, tab/space indents, brace location, alignments,
binpacking etc... were wildly inconsistant in this file.  It
made it hard to update without introducing unintentded changes
and difficult to read.

I've cleaned up most of the whitespace inconsistancy.  I can't say this
is absolutelty complete, but should be the majority.  Some deviation
may occur in the intrest of readability, and I expect future changes to
favor readabilty as well.

As I'm starting a process to clean-up this dissector and improve on
remnants left over from its coversion to a built-in from a plug-in, I'm
starting with a whitespace-only commit that funcional changes can base
from.

Change-Id: I2eb0aca8860c2e91daf103b760461b928873a92b
Reviewed-on: https://code.wireshark.org/review/36003
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
Petri-Dish: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>