WireGuard: fix null pointer crash
It was assumed that the WireGuard dissector is always called on the
first pass before the second pass. That might not be the case when the
heuristics dissector sets a conversation dissector later in the stream.
Be prepared to handle this case. Do not simply abort, the previous
packets may be valid data messages.

Bug: 16394
Change-Id: Id5bf38c07f4d1bffd4b372e92d9a8784e094829a
Reviewed-on: https://code.wireshark.org/review/36148
Petri-Dish: Peter Wu <peter@lekensteyn.nl>
Tested-by: Petri Dish Buildbot
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
(cherry picked from commit b2ee7a2876c26b71f50cd0808c89af6a8483f007)
Reviewed-on: https://code.wireshark.org/review/36159
Reviewed-by: Anders Broman <a.broman58@gmail.com>