FTDI MPSSE: Dissect setting the clock divisor
This patch dissects the clock divisorset command. For the FT2232D it
will also calculate the clock based on the divisor.

Calculating the clock for newer chips requires more context that what we
currently have. The master clock is 60MHz by default but can be
configured to 12MHz for compatibility with older chips. We need to track
the "Divide by 5" commands (0x8A/0x8B) to be able to claculate the
clock.

Ping-Bug: 11743
Change-Id: Ica3acfa97b4db38c2f28bb14cb1e0576d9b5139e
Signed-off-by: Filipe Laíns <lains@archlinux.org>
Reviewed-on: https://code.wireshark.org/review/36167
Petri-Dish: Tomasz Moń <desowin@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Tomasz Moń <desowin@gmail.com>