goose: test field is renamed to simulation
The GOOSE dissector included Wireshark is using the GOOSE message
definition that was described in an older version of the IEC 61850
standard. To be precise, the current field names in Wireshark
corresponds to the Edition 1 of the IEC 61850 standard.
This GOOSE message definition is defined in Table 23 of the
IEC 61850-7-2:2003.
In this release (IEC 61850-7-2:2003) there is a field name called: Test.

However, in the new version of the IEC 61850, the Edition 2 of the
IEC 61850 standard, the Test field was renamed to Simulation.
To be precise, the new GOOSE message definition is defined
in Table 43 of the IEC 61850-7-2:2010.

Bug: 16402
Change-Id: I407987d7f4564b5bafa0c9217756c3f9f23918f6
Reviewed-on: https://code.wireshark.org/review/36175
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>