Qt: Don't hijack Ctrl+C for copy packet list entry as plain text
Don't hijack Ctrl+C / Cmd+C for copy Packet List entry as plain text
in the main window because this will affect using this shortcut in
other dialogs.

Ctrl+C will now copy the selected row in the Packet List (without
the headers), or the selected entry in the Packet Details, and it
also works correctly in other dialogs.

This was introduced in gb3f240dbf8.

Change-Id: I99d87f49c436a6dbf98e51d638a09c15a18bcc72
Reviewed-on: https://code.wireshark.org/review/36280
Petri-Dish: Stig Bjørlykke <stig@bjorlykke.org>
Reviewed-by: Anders Broman <a.broman58@gmail.com>