afs: fix backup & butc RPC confusion
The AFS dissector was erroneously attempting to dissect the RPCs for the
Backup Tape Controller (BUTC) on the BACKUP port 7021.  A different set
of RPCs for the Backup Database Server are served on the backup port
7021.

To address this confusion:
- Rename the existing "backup" RPC table to "butc", and dissect it on
the correct port.  This port may be in a range from 7025 up to the
maximum 65535.
- Create a new "backup" RPC table with the correct RPC names for those
exported on the 7021 port.
- Make other adjustments as needed to properly distinguish and handle
both "backup" and "butc" requests and replies.

With these changes, the BKUP operations in the Wireshark afs.cap sample
capture are now dissected correctly.

Change-Id: I906787889e10cad307d9a6fa610c3e35f2b69184
Reviewed-on: https://code.wireshark.org/review/36300
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>