gitlab-ci: upgrade to Clang 10, drop special -Wframe-larger-than case
Requires an updated wireshark/wireshark-ubuntu-dev image:
https://github.com/wireshark/wireshark-ubuntu-dev-docker/pull/2

Remove -Wframe-larger-than while at it. The default size in our CMake
config is 32k. Clang should not significantly go over it. If so, then it
has to be solved there, and not in the Gitlab config.

Change-Id: I3891fcbd9dec8e5a4597404aa8131f28a1755a02
Reviewed-on: https://code.wireshark.org/review/36369
Reviewed-by: Peter Wu <peter@lekensteyn.nl>
(cherry picked from commit 7579d9844839d34ac5fcfcc6e3037a6ddc381583)
Reviewed-on: https://code.wireshark.org/review/36570