TLS: Limit the number of DNs in Certificate Request messages
When the Distinguished Names list is large in the Certificate Request it can
trigger the "Dissector bug warning". Having a huge list in the tree pane is
not useful anyway so this change limits the amount of DNs added to the tree,
preventing this fault condition.
It is indicated by adding a "[Tree view truncated]" item.

A side effect is that a handshake is no longer incorrectly flagged as 'resumed'
because the Server Hello Done in the same packet is now dissected.

Bug: 16202
Change-Id: Ib315940dcabc2d6b31cf3562354214158ea545a5
Reviewed-on: https://code.wireshark.org/review/36314
Reviewed-by: Peter Wu <peter@lekensteyn.nl>
Petri-Dish: Peter Wu <peter@lekensteyn.nl>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>
(cherry picked from commit 914bb159f781c2f3a788fb68a3bacd3f54e24179)
Reviewed-on: https://code.wireshark.org/review/36351
Reviewed-by: André Luyer <andre@luyer.nl>