Fix formatting of the start time and elapsed time.
There's no

    QString arg(uint a, int fieldWidth = 0, QChar fillChar = QLatin1Char(' '))
        const

version of the arg method of QString, there's only

    QString arg(uint a, int fieldWidth = 0, int base = 10,
        QChar fillChar = Latin1Char(' ')) const

so if you don't pass the base argument, it turns the QChar into an int
(presumably using the unicode method, so that ends up being 0x20 or 32),
passes it as the base argument (so it does the conversion base-32 -
that's <= 36, so it's a valid value for the base argument), and defaults
the fillChar argument to space.

Add 10 as the base argument, so it behaves correctly.

Bug: 16429
Change-Id: If4872d6d55aa5d9a7489219622d4190827e65d34
Reviewed-on: https://code.wireshark.org/review/36337
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>