USB HID: refactor usage page identification
The current approach to identify the the usage page and usage page usages
is not great. It joins both the usage pages and the usages which are
supposed to be individual values. The HID usage tables declarations also
become harder to read as they hold the constructed value, not the value
from the spec.

This patch changes the HID usage table declartions from range_string to
value_string, making them less bulky and easier to read, and splits out
the usage page from the usages.

Change-Id: I018406267fae43683d68c1349049e2c7c26f8ad6
Signed-off-by: Filipe Laíns <lains@archlinux.org>
Reviewed-on: https://code.wireshark.org/review/36362
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>