USB HID: make the dissection tree look as in the spec
Currently, the top dissection tree looks like this:

Global item (Usage Page)
Local item (Usage)
Main item (Collection)

This approach is not great: it provides little relevant information. The
item type (Main,Global,Local) is most of the time not relevant, it is
used to differentiate items within the spec and is not useful to the end
user. The item name (Usage Page, Usage, Collection, etc.) and its value
is much more relevant to the end user. We show the name, which is great,
but we don't show the value, forcing the user to open the subtree to
access it.

The report descriptors described in the spec are annotated like this:

Usage Page (Generic Desktop)
Usage (Mouse)
Collection (Application)

This patch makes the dissection tree look like that. All the information
is still present in the tree but we put the information most people are
actually interested in in the top level item.

Change-Id: Idbfb0f137bc82b5e69787ee524585d18eb697340
Signed-off-by: Filipe Laíns <lains@archlinux.org>
Reviewed-on: https://code.wireshark.org/review/36374
Petri-Dish: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Tomasz Moń <desowin@gmail.com>