packet-dis: fix dissection of Entity State Update PDUs
The dissector for Distributed Interactive Simulation reports malformed packets
for Entity State Update PDUs because the offset for the Number of Variable
Records field is off-by-one.  The Padding field dissection is also off-by-one.

This change fixes the problem by not incrementing the offset returned by the
parseField_Entity function which is already past the Entity ID field.  The
offset must be incremented after dissecting the Number of Variable Records
field to get to the start of the Entity Linear Velocity.

Change-Id: I942123be7a1d4e4e4fc587f757cb030a3daf8ef2
Reviewed-on: https://code.wireshark.org/review/36420
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>