Add support for more monitor opcodes.
Support more of what current Linux kernels generate.  We don't yet
dissect control event payloads.

While we're at it:

Opcodes are defined in the kernel source in decimal; define and show
them as decimal.

Use #defines for a bunch of enumerated values, such as opcodes

Show the opcode's numerical value in the Info column if it's not one we
know about.

Change-Id: I915981a46d1a4a544a5e036e69d9a9de8cb8cb30
Reviewed-on: https://code.wireshark.org/review/36487
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>