s_addr -> src_addr, d_addr -> dst_addr.
On at least some platforms (at least some versions of the Windows SDK,
and netinet/in.h on Illumos and thus possibly Solaris, and possibly
others), struct in_addr's only member is a union, with some members of
the union being substructures with fields corresponding to the pre-CIDR
parts of an IPv4 address, with s_addr #defined to be the member of the
union corresponding to the IPv4 address as a 32-bit integral value.

That's presumably maintained for source compatibility with really old
programs written before CIDR, on platforms where they worry about that.

The *BSDs don't care, and macOS came out post-CIDR and never supportede
the pre-CIDR code in the first place, so they don't do that; the same
probably applies to Linux.

This causes problems when s_addr is used as a variable name.

In addition, inside Wireshark we tend to use "src" rather than just "s"
as a prefix to refer to source addresses and ports - and "dst" or "dest"
rather than just "d" as a prefix to refer to destination addresses and
ports.

So just go with that.

Change-Id: I6010d5ea9697d2333597c06756fa393a5898364e
Reviewed-on: https://code.wireshark.org/review/36504
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>