rlcmac: Don't pass array element to CSN1 descriptors
This way the macros can be used to access the arrays themselves and
calculate its static size to enable validation lateri on.

In the case of Packet_Access_Reject_t, modify the description to use a
M_REC_TARRAY_1 object to get rid of access to 2nd element. The new
description is the correct one, since the first element is mandatory
according to TS 44.060 Table 11.2.1.

Port from osmo-pcu.git 81b40cbaf3070f70954663f68375100128bdc77e.

Change-Id: I48fc257305af799beb6a48ffc6d11b73026bff51
Reviewed-on: https://code.wireshark.org/review/36571
Reviewed-by: Harald Welte <laforge@gnumonks.org>
Reviewed-by: Vadim Yanitskiy <axilirator@gmail.com>
Reviewed-by: Pascal Quantin <pascal@wireshark.org>