github: add cmake options workflow.
This workflow will test the several options available in cmake,
by not using the default value.

The workflow runs once a day, instead on push, to spot problems
that unlikely happen.

The compilation without pcap has been removed from other CIs,
since it is included in this one and that will spare CI cycles.

Change-Id: I796a1ac1879fe85c66d9518207c7053531204c11
Reviewed-on: https://code.wireshark.org/review/36608
Reviewed-by: Peter Wu <peter@lekensteyn.nl>